backend/requirements.txt
fastapi
pydantic
ollama
nest_asyncio
uvicorn
pymongo
flask
requests
google-generativeai